test: split PDF caption-bounded text ownership coverage
